pin support for nan_box

+++ cyarg/builtin.c
@@ -344,7 +344,7 @@ bool pinBuiltin(ObjRoutine* routineContext, int argCount, Value* args, Value* re
         if (vm.pinnedRoutines[i] == NULL) {
             vm.pinnedRoutines[i] = AS_ROUTINE(args[0]);
 #ifdef CYARG_PICO_TARGET
-            *result = UINTEGER_VAL(vm.pinnedRoutineHandlers[i]);
+            *result = UINTEGER_VAL((uintptr_t)vm.pinnedRoutineHandlers[i]);
 #else
             *result = UINTEGER_VAL(i);
 #endif
